BaseS3Object was updated to better protect its metadata information, in particular to make its internal metadata map available only as an immutable object.
Also, S3Object was updated such that an invalid object (ie one without a key) cannot be easily created.

These changes caused wide-ranging cascading code changes, in main part due to the bodgy way in which object's metadata was accessed and set previously.

Overall, this cleanup is a pain but it's worth it to have the better data protection practices enforced.